feat: add custom PluginList component for more consistent UX when choosing plugins
Previously, the initial plugin list implementation had only basic functionality. This meant that the UX was wildly inconsistent, with arrow keys not working. This component re-implements many of the conveniences in List.svelte. In the future, we may somehow extract the duplicate logic, but it's not too much for now.